+++ Patches/ChatCommandPatch.cs
@@ -658,7 +658,7 @@ internal static class ChatCommands
         HudManagerStartPatch.TryResizeUI(scale);
     }
     
-    private static void SelectCommand(PlayerControl player, string text, string[] args)
+    public static void SelectCommand(PlayerControl player, string text, string[] args)
     {
         if (Starspawn.IsDayBreak) return;
 
@@ -900,7 +900,7 @@ internal static class ChatCommands
         }
     }
     
-    private static void ChooseCommand(PlayerControl player, string text, string[] args)
+    public static void ChooseCommand(PlayerControl player, string text, string[] args)
     {
         if (!Main.PlayerStates.TryGetValue(player.PlayerId, out var state) || state.IsDead || state.Role is not Pawn pawn) return;
         
@@ -916,7 +916,7 @@ internal static class ChatCommands
         MeetingManager.SendCommandUsedMessage(args[0]);
     }
 
-    private static void ForgeCommand(PlayerControl player, string text, string[] args)
+    public static void ForgeCommand(PlayerControl player, string text, string[] args)
     {
         if (Starspawn.IsDayBreak) return;
 
@@ -1024,7 +1024,7 @@ internal static class ChatCommands
             return;
         }
 
-        Utils.SendMessage(GetString($"8BallResponse.{IRandom.Instance.Next(Options.EightballCommandIndexes.GetInt())}"), player.IsAlive() ? byte.MaxValue : player.PlayerId, GetString("8BallResponseTitle"));
+        Utils.SendMessage(GetString($"8BallResponse.{IRandom.Instance.Next(20)}"), player.IsAlive() ? byte.MaxValue : player.PlayerId, GetString("8BallResponseTitle"));
     }
 
     public static void GameModePollCommand(PlayerControl player, string text, string[] args)
@@ -2194,7 +2194,7 @@ internal static class ChatCommands
         AmongUsClient.Instance.KickPlayer(kickedPlayer.OwnerId, args[0] == "/ban");
     }
 
-    private static void CheckCommand(PlayerControl player, string text, string[] args)
+    public static void CheckCommand(PlayerControl player, string text, string[] args)
     {
         if (Starspawn.IsDayBreak) return;
 
@@ -2683,7 +2683,9 @@ internal static class ChatCommands
     private static void RCommand(PlayerControl player, string text, string[] args)
     {
         string subArgs = text.Remove(0, 2);
-        byte to = player.AmOwner && (Input.GetKey(KeyCode.LeftControl) || Input.GetKey(KeyCode.RightControl)) ? byte.MaxValue : player.PlayerId;
+        byte to = player.AmOwner && ((Input.GetKey(KeyCode.LeftControl) || Input.GetKey(KeyCode.RightControl)) || ClientControlGUI.BroadcastRoleInfo)
+            ? byte.MaxValue : player.PlayerId;
+            ClientControlGUI.BroadcastRoleInfo = false;
         SendRolesInfo(subArgs, to);
     }