+++ Roles/Standard/Impostor/Miscellaneous/Loner.cs
@@ -1,5 +1,8 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using EHR.Modules;
+using UnityEngine;
+using static EHR.Translator;
 
 namespace EHR.Roles;
 
@@ -46,12 +49,54 @@ public class Loner : RoleBase
     public override void AfterMeetingTasks()
     {
         var pc = PickedPlayer.GetPlayer();
-        
+
         if (!Done && PickedPlayer != byte.MaxValue && PickedRole != CustomRoles.Crewmate && pc != null)
         {
             Done = true;
             pc.RpcSetCustomRole(PickedRole);
             pc.RpcChangeRoleBasis(PickedRole);
         }
+    }
+
+    public static void CreateLonerButton(MeetingHud __instance)
+    {
+        foreach (PlayerVoteArea pva in __instance.playerStates)
+        {
+            PlayerControl pc = Utils.GetPlayerById(pva.TargetPlayerId);
+            if (!pc || !pc.IsAlive()) continue;
+
+            GameObject template = pva.Buttons.transform.Find("CancelButton").gameObject;
+            GameObject targetBox = Object.Instantiate(template, pva.transform);
+            targetBox.name = "ShootButton";
+            targetBox.transform.localPosition = new(-0.35f, 0.03f, -1.31f);
+            var renderer = targetBox.GetComponent<SpriteRenderer>();
+            renderer.sprite = Utils.LoadSprite("EHR.Resources.Images.Skills.Sidekick.png", 160f);
+            var button = targetBox.GetComponent<PassiveButton>();
+            button.OnClick.RemoveAllListeners();
+            button.OnClick.AddListener((Action)(() => GuessManager.GuesserOnClick(pva.TargetPlayerId, __instance, true)));
+        }
+    }
+
+    //[HarmonyPatch(typeof(MeetingHud), nameof(MeetingHud.Start))]
+    public static class StartMeetingPatch
+    {
+        public static void Postfix(MeetingHud __instance)
+        {
+            if (PlayerControl.LocalPlayer.Is(CustomRoles.Loner) && PlayerControl.LocalPlayer.IsAlive())
+                CreateLonerButton(__instance);
+        }
+    }
+
+    public static void ProcessGuesserUI(byte playerId, CustomRoles role)
+    {
+        PlayerControl pc = Utils.GetPlayerById(playerId);
+        if (pc == null || !pc.IsAlive() || Starspawn.IsDayBreak) return;
+
+        var command = $"/select {playerId} {GetString(role.ToString())}";
+
+        if (AmongUsClient.Instance.AmHost)
+            ChatCommands.SelectCommand(PlayerControl.LocalPlayer, command, command.Split(' '));
+        else
+            ChatCommands.RequestCommandProcessingFromHost(command, "Select");
     }
 }