+++ Roles/Standard/Neutral/Killing/Jackal.cs
@@ -171,12 +171,12 @@ public class Jackal : RoleBase
 
         Main.ResetCamPlayerList.Add(target.PlayerId);
 
-        hasValue |= sender.Notify(killer, CustomRoles.Jackal.ColoredTextByRole(GetString("GangsterSuccessfullyRecruited")), setName: false);
+        hasValue |= CustomRpcSenderExtensions.Notify(ref sender, killer, CustomRoles.Jackal.ColoredTextByRole(GetString("GangsterSuccessfullyRecruited")), setName: false);
         hasValue |= sender.SetKillCooldown(killer, 3f);
         hasValue |= sender.NotifyRolesSpecific(killer, target, out sender, out bool cleared);
         if (cleared) hasValue = false;
 
-        hasValue |= sender.Notify(target, CustomRoles.Jackal.ColoredTextByRole(GetString("BeRecruitedByJackal")), setName: false);
+        hasValue |= CustomRpcSenderExtensions.Notify(ref sender, target, CustomRoles.Jackal.ColoredTextByRole(GetString("BeRecruitedByJackal")), setName: false);
         hasValue |= sender.RpcGuardAndKill(target, killer);
         hasValue |= sender.RpcGuardAndKill(target, target);
         hasValue |= sender.NotifyRolesSpecific(target, killer, out sender, out cleared);