+++ Roles/Standard/Neutral/Killing/Necromancer.cs
@@ -221,12 +221,12 @@ internal class Deathknight : RoleBase
             var sender = CustomRpcSender.Create("Deathknight.OnCheckMurder", SendOption.Reliable);
             var hasValue = false;
 
-            hasValue |= sender.Notify(killer, CustomRoles.Necromancer.ColoredTextByRole(GetString("DeathknightRecruitedPlayer")));
+            hasValue |= CustomRpcSenderExtensions.Notify(ref sender, killer, CustomRoles.Necromancer.ColoredTextByRole(GetString("DeathknightRecruitedPlayer")));
             hasValue |= sender.SetKillCooldown(killer);
             hasValue |= sender.NotifyRolesSpecific(killer, target, out sender, out bool cleared);
             if (cleared) hasValue = false;
 
-            hasValue |= sender.Notify(target, CustomRoles.Necromancer.ColoredTextByRole(GetString("RecruitedByDeathknight")));
+            hasValue |= CustomRpcSenderExtensions.Notify(ref sender, target, CustomRoles.Necromancer.ColoredTextByRole(GetString("RecruitedByDeathknight")));
             hasValue |= sender.RpcGuardAndKill(target, killer);
             hasValue |= sender.RpcGuardAndKill(target, target);
             hasValue |= sender.NotifyRolesSpecific(target, killer, out sender, out cleared);